% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff2n.R
\name{ff2n}
\alias{ff2n}
\title{A function generates an index matrix}
\usage{
ff2n(s)
}
\arguments{
\item{s}{the dimension of the matrix, \code{s} shouldn't be large}
}
\value{
a \code{2^s*s} index matrix filled with 0 and 1
}
\description{
A function generates an index matrix
}
\examples{
\dontrun{
Index <- ff2n(3)
print(Index)
}
}
