% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff2nC}
\alias{ff2nC}
\title{A function generates an index matrix}
\usage{
ff2nC(p)
}
\arguments{
\item{p}{the dimension of the matrix, \code{p} shouldn't be large}
}
\value{
a \code{2^s*s} index matrix filled with 0 and 1
}
\description{
A function generates an index matrix
}
\examples{
\dontrun{
Index <- ff2n(3)
print(Index)
}
}
