% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictMA.R
\name{PredictMA}
\alias{PredictMA}
\title{A prediction function after model averaging}
\usage{
PredictMA(Data, MA, ynew = FALSE)
}
\arguments{
\item{Data}{New data contain explanatory variables in whole model}

\item{MA}{Result of function `ModelAve()`}

\item{ynew}{Whether a new \code{y} is in \code{Data} or is given as a vector. If `ynew=TRUE`, \code{Data} should contain the explanatory vector,
and if `ynew` is a vector with same size as \code{Data}, if `ynew=FALSE`, there is no new \code{y}.}
}
\value{
Prediction of data
}
\description{
Predicting with new data
}
\examples{
\dontrun{
## simulation data
n <- 10
p <-3
X <- data.frame(matrix(rnorm(n*p), n, p))
beta <- c(1, 2, 3)
y <- X \%*\% beta
Data <- cbind(y, X)
subset <- matrix(c(1,0,0,0,1,0),ncol=p)
get_MA<-MA(y~., Data, method="MMA", Subset = subset)
PredictMA(data.frame(matrix(rnorm(5*p),ncol=p)),get_MA)
}
}
