% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAve.R
\name{ModelAve}
\alias{ModelAve}
\title{A model averaging function}
\usage{
ModelAve(
  Formula,
  Data,
  nested = FALSE,
  method,
  Subset = NULL,
  equalweight = FALSE
)
}
\arguments{
\item{Formula}{formula like `y~.` or `y~X1+X2` which mean whole model}

\item{Data}{Data set including response variable and explanatory variables, the first column is response variable}

\item{nested}{Whether using nested model, `default=FALSE`}

\item{method}{One of "MMA" and "JMA"}

\item{Subset}{An index matrix of n*p with 0-1 element, \code{n} is model number and \code{p} is whole explanatory variables number}

\item{equalweight}{Whether use equal weight, `default=FALSE`}
}
\value{
The optimal weight, the coefficients of all models
}
\description{
Model averaging function including MMA and JMA
}
\examples{
\dontrun{
## simulation data
n <- 10
p <-3
X <- data.frame(matrix(rnorm(n*p), n, p))
beta <- c(1, 2, 3)
y <- X \%*\% beta
Data <- cbind(y, X)
subset <- matrix(c(1,0,0,0,1,0),ncol=p)
MA(y~., Data, method="MMA", Subset = subset)
}
}
